backend/requirements.txt
tensorflow
numpy
pillow
flask
flask-cors